+++ test/EndToEnd/RuleSet/Php53Test.php
@@ -1,0 +1,23 @@
+<?php
+
+declare(strict_types=1);
+
+/**
+ * Copyright (c) 2019-2023 Andreas Möller
+ *
+ * For the full copyright and license information, please view
+ * the LICENSE.md file that was distributed with this source code.
+ *
+ * @see https://github.com/ergebnis/php-cs-fixer-config
+ */
+
+namespace Ergebnis\PhpCsFixer\Config\Test\EndToEnd\RuleSet;
+
+/**
+ * @coversNothing
+ *
+ * @requires PHP 5.3
+ */
+final class Php53Test extends AbstractRuleSetTestCase
+{
+}